New standard modals. Create views replaced by modals
